requirements.txt
streamlit
langchain-community
langchain-core
chromadb==0.4.18
sentence-transformers
pypdf
pydantic==1.10.13